network modal fix
