+++ components/special/NoConnectionModal.tsx
@@ -1,3 +1,4 @@
+import NetInfo from '@react-native-community/netinfo'
 import { useMutation } from '@tanstack/react-query'
 import { useEffect, useState } from 'react'
 
@@ -8,9 +9,10 @@ import { useTranslation } from '@/hooks/useTranslation'
 import { clientApi } from '@/modules/backend/client-api'
 import { useAxiosResponseInterceptors } from '@/modules/backend/hooks/useAxiosResponseInterceptors'
 
-const AxiosConnection = () => {
-  const t = useTranslation('AxiosConnection')
+const NoConnectionModal = () => {
+  const t = useTranslation('NoConnection')
 
+  const [isConnected, setIsConnected] = useState(true)
   const [serverConnectionError, setServerConnectionError] = useState(false)
   useAxiosResponseInterceptors(setServerConnectionError)
 
@@ -18,9 +20,24 @@ const AxiosConnection = () => {
     mutationFn: () => {
       return clientApi.appControllerHealth()
     },
-    onSuccess: () => setServerConnectionError(false),
+    onSuccess: () => {
+      setIsConnected(true)
+      setServerConnectionError(false)
+    },
     onError: () => setServerConnectionError(true),
   })
+
+  useEffect(() => {
+    const unsubscribe = NetInfo.addEventListener((networkState) => {
+      if (!networkState.isConnected) {
+        setIsConnected(false)
+      }
+    })
+
+    return () => {
+      unsubscribe()
+    }
+  }, [])
 
   useEffect(() => {
     healthCheck.mutate()
@@ -28,13 +45,15 @@ const AxiosConnection = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
+  const translationName = isConnected ? 'server' : 'network'
+
   return (
     <Modal visible={serverConnectionError}>
       <ModalContentWithActions
         customAvatarComponent={<AvatarCircleNetworkOff />}
-        title={t('title')}
+        title={t(`${translationName}.title`)}
         isLoading={healthCheck.isPending}
-        text={t('text')}
+        text={t(`${translationName}.text`)}
         primaryActionLabel={t('primaryActionLabel')}
         primaryActionOnPress={healthCheck.mutate}
       />
@@ -42,4 +61,4 @@ const AxiosConnection = () => {
   )
 }
 
-export default AxiosConnection
+export default NoConnectionModal
